backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
mysql-connector-python==8.2.0
pandas==2.1.4
numpy==1.26.3
scikit-learn==1.3.2
xgboost==2.0.3
lightgbm==4.1.0
shap==0.43.0
joblib==1.3.2
paho-mqtt==1.6.1
kafka-python==2.0.2
redis==5.0.1
websockets==12.0
aiohttp==3.9.1
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
cryptography==41.0.7
sqlalchemy-utils==0.41.1
alembic==1.13.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
